% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedPlots.R
\name{combinedPlots}
\alias{combinedPlots}
\title{Generate Combined Plots (Internal)}
\usage{
combinedPlots(
  ExperimentData,
  dt_final,
  dt_finalSummary,
  font,
  divisions,
  pValues
)
}
\arguments{
\item{ExperimentData}{An S4 object containing experiment metadata.}

\item{font}{A character string variable determining the font style of the produced plots.}

\item{divisions}{A list of grouping columns used for facetting plots.}

\item{pValues}{A TRUE/FALSE vector for if combined plots will display p values for 2-condition overlays.}

\item{dt_curated_final}{A \code{data.table} containing curated sleep data with columns such as \code{id} and \code{asleep}.}

\item{summary_dt_final}{A \code{data.table} containing summary statistics with columns including \code{Light}, \code{Environment},
\code{Genotype}, \code{Treatment}, \code{Sex} and various sleep metrics.}
}
\value{
None. Plots are saved as PDF files.
}
\description{
Creates combined plots for each unique combination of \code{Light}, \code{Environment}, \code{Treatment}, \code{Sex}, and \code{Genotype} in the dataset.
Generates overlay sleep plots and sleep duration plots, saving each combination as a PDF file.
}
\details{
The function iterates through all unique combinations of \code{Light}, \code{Environment}, \code{Treatment}, \code{Sex}, and \code{Genotype}.
For each combination:
\itemize{
\item An overlay sleep plot is created using \code{ggetho}.
\item Multiple sleep duration plots (e.g., total sleep, daytime sleep, nighttime sleep) are generated
using a helper function, \code{create_sleeptime_plot}.
\item Plots are combined into a single figure using \code{cowplot::plot_grid}.
\item The combined figure is saved as a PDF file, with the filename reflecting the combination of \code{Light},
\code{Environment}, and \code{Genotype}.
}

The function dynamically adjusts plot widths based on the number of unique possibilities of the first divisions entry (a parameter).
}
\keyword{internal}
