% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate Summary Tables (Internal)}
\usage{
concatenate(dir, all_tables, patternname)
}
\arguments{
\item{dir}{A character string specifying the directory containing the summary files.}

\item{all_tables}{A list to store and update with the loaded data tables.}

\item{patternname}{A character string pattern to match the summary CSV files.}
}
\value{
An updated list of data tables with the newly loaded data appended.
}
\description{
An internal helper function that reads summary CSV files matching a specified pattern from a directory and appends them to a list of tables.
}
\keyword{internal}
