% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatGenotypePlots.R
\name{concatGenotypePlots}
\alias{concatGenotypePlots}
\title{Generate Combined Genotype Plots (Internal)}
\usage{
concatGenotypePlots(
  combined_sleepdata,
  combined_sleepmeta,
  summary_dt_final,
  control,
  font
)
}
\arguments{
\item{combined_sleepdata}{A \code{data.table} summarized accross all batches containing curated sleep data with columns such as \code{id} and \code{asleep}.}

\item{combined_sleepmeta}{A \code{data.table} summarized accross all batches containing the metadata associated with combined_sleepdata.}

\item{summary_dt_final}{A \code{data.table} containing summary statistics with columns including \code{Light}, \code{Environment},
\code{Genotype}, \code{Treatment}, \code{Sex} and various sleep metrics.}

\item{control}{A character string specifying the control from \code{divisions[1]}.}

\item{font}{A character string variable determining the font style of the produced plots.}
}
\value{
None. Plots are saved as PDF files.
}
\description{
Creates combined plots for each unique combination of \code{Light}, \code{Environment}, and \code{Genotype} in the dataset.
Generates overlay sleep plots and sleep duration plots, saving each combination as a PDF file.
}
\details{
The function iterates through all unique combinations of \code{Light}, \code{Environment}, \code{Treatment}, \code{Sex}, and \code{Genotype}.
For each combination:
\itemize{
\item An overlay sleep plot is created using \code{ggetho}.
\item Multiple sleep duration plots (e.g., total sleep, daytime sleep, nighttime sleep) are generated
using a helper function, \code{create_sleeptime_plot}.
\item Plots are combined into a single figure using \code{cowplot::plot_grid}.
\item The combined figure is saved as a PDF file, with the filename reflecting the combination of \code{Light},
\code{Environment}, and \code{Genotype}.
}
}
\keyword{internal}
