% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMat.R
\name{corMat}
\alias{corMat}
\title{Correlation Matrix for Sleep Data (Exported)}
\usage{
corMat(Compare1, Compare2)
}
\arguments{
\item{Compare1}{A character string representing the first treatment to compare.}

\item{Compare2}{A character string representing the second treatment to compare.}

\item{combined_data}{A data frame containing combined data from all batches, including variables such as \code{temp}, \code{sex}, \code{treatment}, \code{genotype}, and sleep-related metrics.}
}
\value{
A \code{ggplot} object containing the correlation matrix plot with significance annotations.
The plot is saved as a PDF file named \verb{RelativeBouts<Compare1>.pdf}.
A CSV file with correlation p-values is saved as \verb{statcor<Compare1>.csv}.
}
\description{
Computes the relative sleep changes between two treatments and generates a correlation matrix plot with significance annotations.
}
