% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Setup Environment and Define ExperimentData Class (Internal)}
\usage{
.onLoad(libname, pkgname)
}
\description{
This internal function checks if the \code{ExperimentData} class is already defined.
If not, it defines the class to be used for storing experimental data in downstream functions.
}
\keyword{internal}
