% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDisplay.R
\name{normDisplay}
\alias{normDisplay}
\title{Visualize Changes in Sleep Between}
\usage{
normDisplay(
  x,
  treat = NULL,
  temp = NULL,
  enviro = NULL,
  Lights = NULL,
  geno = NULL,
  font = "plain"
)
}
\arguments{
\item{x}{The name of a parameter in "all_batches_norm_summary.csv" to be used as the x-axis in the plot.}

\item{treat}{A string specifying a Treatment to subset the data by.}

\item{temp}{A string specifying a Temperature condition to subset the data by.}

\item{enviro}{A string specifying a Environment condition to subset the data by.}

\item{Lights}{A string specifying a Light condition to subset the data by.}

\item{geno}{A string specifying a Genotype condition to subset the data by.}

\item{font}{A character string determining the font style of the produced plots. ("plain", "bold", "italic", or "bold.italic")}
}
\value{
None. Saves the plots as PDF files labeled \code{NormalizedSleepLoss...}
}
\description{
This function generates bar plots to visualize the changes in sleep across different Genotypes and Treatments.
It reads a CSV file containing normalized sleep data, calculates the changes in sleep, and then creates plots
for each of the sleep-related metrics (\code{TotalSleepChange}, \code{DayTimeSleepChange}, \code{NightTimeSleepChange}).
Two plots are made showing the normalized sleeploss for the Treatment of interest (\code{treat}).
One plot contains only the \code{control} entry in \code{x} while the other plot contains every entry except \code{control}.
The plots are saved as PDFs labeled, "NormalizedSleepLoss...".
}
\keyword{internal}
