% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPreferences.R
\name{plotPreferences}
\alias{plotPreferences}
\title{Ask the user which plots they want (Internal)}
\usage{
plotPreferences(run = "all")
}
\arguments{
\item{run}{A character string of either "all" or "one" referring to which run is done}
}
\value{
A matrix containing the user's plot preferences, which can be used for downstream analysis.
}
\description{
This function prompts the user to select which plots they do not wish to generate. It presents a series of
Yes/No questions regarding different types of plots, and returns their preferences.
for downstream analysis.
}
\keyword{internal}
