% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processDays.R
\name{processDays}
\alias{processDays}
\title{Process Latency and Bout Lengths for Each Day (Internal)}
\usage{
processDays(numDays, bout_dt, dt, summary_dt_final)
}
\arguments{
\item{numDays}{An integer specifying the number of days for which to process the bout data.}

\item{bout_dt}{A data.table containing bout data, including time (\code{t}) and bout duration (\code{duration}).}

\item{dt}{data table with raw sleep data}

\item{summary_dt_final}{A data.table containing the final summary data, which will be updated with the computed
latency and bout lengths for each day.}
}
\value{
A data.table containing the final summary with additional columns for latency, first bout length, and
latency to the longest bout for each day.
}
\description{
This function processes bout data to calculate the latency to the first bout, the length of the first bout,
and the latency to the longest bout for each animal per day. The function computes these statistics for each
day within a specified number of days and merges the results into a final summary table.
}
\details{
The function processes bout data for each day within the specified range, calculating the latency to the first bout,
the first bout length, and the latency to the longest bout. The results for each day are added to a summary table,
which is then returned after processing all specified days.
}
\keyword{internal}
