% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllBatches.R
\name{runAllBatches}
\alias{runAllBatches}
\title{Run All Batches Analysis}
\usage{
runAllBatches(control, font = "plain")
}
\arguments{
\item{control}{A character string specifying the control from one of the variables in the Main.r file.}

\item{font}{A string variable determining the font style of the produced plots.}

\item{parent_dir}{The parent directory where all batch directories are located.}

\item{info}{A data.table containing experimental information.}

\item{divisions}{A list of time divisions for the analysis.}

\item{num_days}{The number of days to consider for the analysis.}
}
\value{
This function does not return a value, but generates and saves two CSV files:
\code{all_batches_norm_summary.csv} and \code{all_batches_summary.csv}.
}
\description{
This function processes all batch directories by sourcing the necessary R files, running
the analysis for each batch, and combining the summary results into a single CSV file for all batches.
It iterates through directories matching the batch pattern, executes relevant R files, and combines
both normalized and general summary statistics from each batch into a final report.
}
\details{
This function performs the following steps for each batch directory:
\enumerate{
\item It iterates over each batch directory matching the batch pattern "Batch\link{0-9}+_\link{0-9}+".
\item For each batch directory, it sources all R files and runs the analysis via the \code{runOneBatch} function.
\item It concatenates the normalized and general summary statistics for each batch into separate CSV files.
\item The final CSV files, \code{all_batches_norm_summary.csv} and \code{all_batches_summary.csv},
are saved in the parent directory, containing combined results for all batches.
}
}
