% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEachBatch.R
\name{runEachBatch}
\alias{runEachBatch}
\title{Run Complete Single Batch Analysis Workflow}
\usage{
runEachBatch(numDays, oneBatch, font, pref, divisions, pValues, incodeinfo)
}
\arguments{
\item{numDays}{A numerical value specifying the number of days to be used in analysis.}

\item{oneBatch}{A character string of the Batch folder to be analyzed.}

\item{font}{A string variable determining the font style of the produced plots.}

\item{pref}{A numeric vector of plot preferences (e.g., pref\link{1} controls actograms, pref\link{6} controls combined plots).}

\item{divisions}{A character vector (length 3) containing the names of the meta-variables used for plot overlay and facetting (e.g., c("Treatment", "Genotype", "Environment")).}

\item{pValues}{A TRUE/FALSE vector for if combined plots will display p values for 2-condition overlays.}

\item{incodeinfo}{A data.table or data.frame containing the batch metadata (e.g., Sex, Genotype, Monitor ID) loaded from the "Main" R file.}
}
\value{
This function does not return a value but performs a series of steps to process the data,
generate plots, and calculate statistics.
}
\description{
This function processes data for each single batch by performing various steps including data cleaning,
generating activity and sleep plots, trimming dead animals, calculating summary statistics, and
generating normalized statistics.
}
\details{
This function executes the following steps:
\enumerate{
\item Initializes an \code{ExperimentData} object containing metadata for the batch.
\item Writes the loading metadata to a CSV file.
\item Generates activity and sleep plots based on the input data.
\item Trims out animals identified as dead using the \code{aliveVsDead} and \code{manualDeadRemoval} functions.
\item Creates a final summary of the bout data and latency statistics.
\item Optionally, generates concatenated genotype plots.
\item Computes summary statistics and normalized statistics for sleep time for all groups.
\item Writes relevant output files, including the final summary and normalized statistics.
}
}
\keyword{Internal}
