% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStatus.R
\name{setStatus}
\alias{setStatus}
\title{Set Status to "notOK" and Exclude for a Specific Region and Monitor}
\usage{
setStatus(info, regionID_targ, monitor_targ)
}
\arguments{
\item{info}{A data.table containing information about regions and monitors, with columns for \code{region_id},
\code{monitor}, and \code{status}.}

\item{regionID_targ}{The ID of the region for which the status should be updated.}

\item{monitor_targ}{The target monitor for which the status should be updated.}
}
\value{
A data.table with the updated status for the specified region and monitor.
}
\description{
This function updates the status of a specific region and monitor combination in the \code{info} data.table.
The status is set to "notOK" for the specified \code{regionID_targ} and \code{monitor_targ}. Subjects with this
status are excluded from further analysis.
}
\details{
This function is typically used by other functions to filter and update the status of subjects
for specific regions and monitors before conducting further analyses.
}
\examples{
# Example usage
info <- data.table::data.table(region_id = c(1, 2, 3),
                               monitor = c("11", "12", "13"),
                               status = c("OK", "OK", "OK"))
updated_info <- setStatus(info, regionID_targ = 1, monitor_targ = "13")
print(updated_info)
}
