% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsSummary.R
\name{statsSummary}
\alias{statsSummary}
\title{Generate Summary Statistics (Internal)}
\usage{
statsSummary(ExperimentData, dt, groups, norm_factor)
}
\arguments{
\item{ExperimentData}{An S4 object containing experimental data, including a \code{Batch} attribute.}

\item{dt}{A data.table containing raw summary statistics to be processed.}

\item{groups}{A character vector specifying the groups for which to calculate summary statistics.}

\item{norm_factor}{A character string indicating the factor used for normalization.}
}
\value{
A data.table containing the summary statistics, which is also saved as a CSV file.
}
\description{
This function calculates summary statistics for specified groups using the \code{GenerateSE} function and writes
the results to a CSV file. The summary includes the raw data processed for each group based on the provided
normalization factor.
}
\keyword{internal}
