% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Compute Summary Statistics (Internal)}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{A data.frame or data.table containing the data to be analyzed.}

\item{measurevar}{A character string specifying the variable for which the summary statistics are calculated.}

\item{groupvars}{A character vector of grouping variables by which to calculate the statistics.}

\item{na.rm}{A logical value indicating whether NA values should be removed before calculations (default: FALSE).}

\item{conf.interval}{A numeric value specifying the confidence interval to compute (default: 0.95).}

\item{.drop}{A logical value to determine whether to drop unused factor levels in the result (default: TRUE).}
}
\value{
A data.frame or data.table containing the summary statistics (N, mean, SD, SE, CI) for each group.
}
\description{
This function calculates summary statistics (N, mean, standard deviation, standard error, and confidence interval)
for a specified measure variable, grouped by one or more grouping variables.
}
\keyword{internal}
