% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLoading.R
\name{writeLoading}
\alias{writeLoading}
\title{Write and Link Loading Metadata (Internal)}
\usage{
writeLoading(ExperimentData)
}
\arguments{
\item{ExperimentData}{An S4 object containing experiment metadata, including the \code{loadinginfo} attribute.}
}
\value{
A linked metadata table for use in downstream analysis.
}
\description{
Writes loading information from the experiment metadata to a CSV file and links the metadata
for further analysis.
}
\keyword{internal}
