% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMatrix.R
\name{corMatrix}
\alias{corMatrix}
\title{Correlation Matrix for Sleep Data}
\usage{
corMatrix(
  condition1 = NULL,
  condition2 = NULL,
  sex = NULL,
  geno = NULL,
  temp = NULL,
  treat = NULL,
  enviro = NULL,
  Lights = NULL,
  font = c("plain", "bold", "italic", "bold.italic")
)
}
\arguments{
\item{condition1}{A string specifying a condition within one of the experimental variables. The difference is calculated as: \code{condition1}-\code{condition2}}

\item{condition2}{A string specifying a condition within the same experimental variable as \code{condition1} that is associated with the difference seen in sleep.}

\item{sex}{A string specifying a Sex condition to subset the data by.}

\item{geno}{A string specifying a Genotype condition to subset the data by.}

\item{temp}{A string specifying a Temperature condition to subset the data by.}

\item{treat}{A string specifying a Treatment to subset the data by.}

\item{enviro}{A string specifying a Environment condition to subset the data by.}

\item{Lights}{A string specifying a Light condition to subset the data by.}

\item{font}{A character string determining the font style of the produced plots. ("plain", "bold", "italic", or "bold.italic")}
}
\value{
None. Plots are saved as PDF files, while p-values are saved as csv files.
}
\description{
Computes the relative sleep changes between two Treatments and generates a correlation matrix plot with significance annotations.
}
\details{
Two \code{ggplot} objects containing the correlation matrix plot with significance annotations for percentage of sleep lost and raw number of minutes lost.
The plots are saved as PDF files.
A CSV file with correlation p-values is saved.
}
