% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corScatter.R
\name{corScatter}
\alias{corScatter}
\title{Display underlying data to the correlation plots}
\usage{
corScatter(
  x = c(NULL, "Sleep_Time_All", "Sleep_Time_L", "Sleep_Time_D", "n_Bouts_L", "n_Bouts_D",
    "mean_Bout_Length_L", "mean_Bout_Length_D"),
  y = c(NULL, "Sleep_Time_All", "Sleep_Time_L", "Sleep_Time_D", "n_Bouts_L", "n_Bouts_D",
    "mean_Bout_Length_L", "mean_Bout_Length_D"),
  condition1 = NULL,
  condition2 = NULL,
  method = c("Diff", "Perc.Change"),
  treat = NULL,
  temp = NULL,
  enviro = NULL,
  sex = NULL,
  lights = NULL,
  geno = NULL,
  aPrioriConditions = NULL,
  aPrioriVariable = c(NULL, "Sex", "Genotype", "Temperature", "Treatment", "Environment",
    "Light"),
  lbf = TRUE,
  formula = NULL,
  font = c("plain", "bold", "italic", "bold.italic")
)
}
\arguments{
\item{x}{A string specifying the x-axis variable of the cluster plot as written in 'all_batches_summary.csv' (e.g., "Sleep_Time_All").}

\item{y}{A string specifying the y-axis variable of the cluster plot as written in 'all_batches_summary.csv' (e.g., "mean_Bout_length_L").}

\item{condition1}{A string specifying the condition used as the reference/numerator (e.g., "Iso") when calculating difference or percent change.}

\item{condition2}{A string specifying the condition used as the comparison/denominator (e.g., "Grp") for difference or percent change.}

\item{method}{The calculation method to use when 'condition1' and 'condition2' are specified. Must be "Diff" (calculates 'condition1\code{-}condition2\verb{) or "Perc.Change" (calculates (}condition1\code{-}condition2\verb{) / }condition2`).}

\item{treat}{A string specifying a Treatment to subset the data by.}

\item{temp}{A string specifying a Temperature condition to subset the data by.}

\item{enviro}{A string specifying a Environment condition to subset the data by.}

\item{sex}{A string specifying a sex condition to subset the data by.}

\item{lights}{A string specifying a Light condition to subset the data by.}

\item{geno}{A string specifying a Genotype condition to subset the data by.}

\item{aPrioriConditions}{A vector of (partial spellings of the) conditions (e.g., c("L1", "L2", "S1", "S2", "CS")) in one of the experimental variables.These are primarily for visualization (coloring) and must exist as a group within the specified aPrioriVariable.}

\item{aPrioriVariable}{A string specifying the variable of the aPrioriConditions. This variable (e.g., "Genotype") provides the column where aPrioriConditions are looked up for plotting.}

\item{lbf}{Add the line of best fit onto the plot.}

\item{formula}{A string defining the independent variables (e.g., "Treatment * Sex") for a linear model (LM). If specified, the plot displays fitted values (residuals removed) instead of raw means. Should be NULL for raw data.}

\item{font}{A character string determining the font style of the produced plots.}
}
\value{
Saves the plot as a PDF file and outputs a CSV file with cluster assignments.
}
\description{
This function reads in a CSV file (\code{all_batches_summary.csv})
The user can provide a list of aPrioriConditions (e.g., Genotype categories) which visually differentiates them from each other in the plots.
}
\details{
This plot best accompanies corMatrix, as it produces the same data, but with the actual points along the line of best fit. This can help with interpreting correlations.
}
