% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDisplay.R
\name{normDisplay}
\alias{normDisplay}
\title{Visualize Changes in Sleep Between Genotypes and Treatments}
\usage{
normDisplay(C1, C2, column_name, WT)
}
\arguments{
\item{C1}{A string specifying the first treatment group.}

\item{C2}{A string specifying the second treatment group.}

\item{column_name}{The name of a column in the normalized summary CSV that will be used for comparison in the data.}

\item{WT}{A string specifying the wild type genotype to be used for comparison.}
}
\value{
None. Saves the plots as PDF files.
}
\description{
This function generates bar plots to visualize the changes in sleep across different genotypes and treatments.
It reads a CSV file containing normalized sleep data, calculates the changes in sleep, and then creates plots
for each of the sleep-related metrics (\code{TotalSleepChange}, \code{DayTimeSleepChange}, \code{NightTimeSleepChange}).
The plots compare the changes between two treatments (C1 and C2) for each genotype, and also include plots
for wild type (WT) genotypes within each treatment group. The resulting plots are saved as PDF files.
}
\details{
The function processes normalized sleep data from a CSV file ("all_batches_norm_summary.csv") and calculates the
change in sleep for each treatment group (\code{C1}, \code{C2}) and genotype, including wild type (WT). The changes in sleep
are calculated for three metrics: \code{TotalSleepChange}, \code{DayTimeSleepChange}, and \code{NightTimeSleepChange}.

For each sleep metric, the function generates bar plots comparing the changes in sleep between the two treatments
(C1 and C2) for each genotype, as well as for wild type genotypes within each treatment group. The plots are saved as
PDF files. The number of batch names is used to adjust the width of the plots, and each plot is formatted to show the
mean change in sleep with customized themes and labels.
}
\keyword{internal}
