% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedDisplay.R
\name{rankedDisplay}
\alias{rankedDisplay}
\title{Visualize Changes in Sleep Between}
\usage{
rankedDisplay(
  x,
  control = NULL,
  condition1 = NULL,
  condition2 = NULL,
  treat = NULL,
  temp = NULL,
  enviro = NULL,
  sex = NULL,
  Lights = NULL,
  geno = NULL,
  ranking = NULL,
  font = "plain",
  limits = c(-100, 1500)
)
}
\arguments{
\item{x}{The name of a parameter in "all_batches_norm_summary.csv" to be used as the x-axis in the plot.}

\item{control}{The name of the control within the parameter x}

\item{condition1}{A string specifying a condition within one of the experimental parameters. The percent difference is calculated as: (\code{condition1}-\code{condition2}) / \code{condition2}}

\item{condition2}{A string specifying a condition within the same experimental parameter as \code{condition1} that is associated with the percent difference seen in sleep.}

\item{treat}{A string specifying a Treatment to subset the data by.}

\item{temp}{A string specifying a Temperature condition to subset the data by.}

\item{enviro}{A string specifying a Environment condition to subset the data by.}

\item{sex}{A string specifying a sex condition to subset the data by.}

\item{Lights}{A string specifying a Light condition to subset the data by.}

\item{geno}{A string specifying a Genotype condition to subset the data by.}

\item{ranking}{A tibble of one column containing the order of conditions displayed in the plot from parameter x}

\item{font}{A character string determining the font style of the produced plots. ("plain", "bold", "italic", or "bold.italic")}

\item{limits}{A list of two numerics to be the upper and lower limits on the plot. Default is c(-100, 1500)}
}
\value{
conditional. If the ranking is NULL (default), then the ranking used in the plots will be returned to be used as the ranking in subsequent rankedDisplay plots. Otherwise, no return. Saves the plots as a PDF file labeled \code{RankedSleep...}
}
\description{
This function generates bar plots to visualize the changes in sleep across different Genotypes and Treatments.
It reads a CSV file containing normalized sleep data, calculates the changes in sleep, and then creates plots
for each of the sleep-related metrics (\code{TotalSleepChange}, \code{DayTimeSleepChange}, \code{NightTimeSleepChange}).
Two plots are made showing the normalized sleeploss for the Treatment of interest (\code{treat}).
One plot contains only the \code{control} entry in \code{x} while the other plot contains every entry except \code{control}.
The plots are saved as PDFs labeled, "NormalizedSleepLoss...".
}
\keyword{internal}
