% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAllBatches.R
\name{runAllBatches}
\alias{runAllBatches}
\title{Run All Batches' Analyses}
\usage{
runAllBatches(
  numDays = 2,
  overlayVar,
  rowVar,
  columnVar,
  plotSelection = "All",
  font = "plain",
  pValues = FALSE
)
}
\arguments{
\item{numDays}{A numerical value specifying the number of days to be used in analysis.}

\item{overlayVar}{A character string specifying which variable to overlay and color plots by Default is "Treatment".}

\item{rowVar}{A character string specifying which variable to facet rows in plots by. Default is "Genotype".}

\item{columnVar}{A character string specifying which variable to facet columns in plots by. Default is "Environment".}

\item{plotSelection}{A character string specifying if the user wants all possible plots, no optional plots, or select specific plot outputs.}

\item{font}{A string variable determining the font style of the produced plots.}

\item{pValues}{A TRUE/FALSE vector for if combined plots will display p values for 2-condition overlays.}
}
\value{
This function does not return a value, but generates and saves two CSV files:
\code{all_batches_norm_summary.csv} and \code{all_batches_summary.csv}.
}
\description{
This function processes all batch directories by sourcing the necessary R files, running
the analysis for each batch, and combining the summary results into a single CSV file for all batches.
It iterates through directories matching the batch pattern, executes relevant R files, and combines
both normalized and general summary statistics from each batch into a final report.
}
\details{
This function performs the following steps for each batch directory:
\enumerate{
\item It iterates over each batch directory matching the batch pattern "Batch...".
\item For each batch directory, it sources all R files and runs the analysis via the \code{runOneBatch} function.
\item It concatenates the normalized and general summary statistics for each batch into separate CSV files.
\item The final CSV files, \code{all_batches_norm_summary.csv} and \code{all_batches_summary.csv},
are saved in the parent directory, containing combined results for all batches.
}
}
