% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEachBatch.R
\name{runEachBatch}
\alias{runEachBatch}
\title{Run the Each Batch Analysis (Internal)}
\usage{
runEachBatch(control, num_days, oneBatch, font = "plain", pref)
}
\arguments{
\item{control}{A character string specifying the control.}

\item{num_days}{A numerical value specifying the number of days to be used in analysis.}

\item{oneBatch}{A character string of the Batch folder to be analyzed.}

\item{font}{A string variable determining the font style of the produced plots.}

\item{pref}{A vector of preferences for generating specific plots (e.g., whether to generate concatenated plots).}
}
\value{
This function does not return a value but performs a series of steps to process the data,
generate plots, and calculate statistics.
}
\description{
This function processes data for each single batch by performing various steps including data cleaning,
generating activity and sleep plots, trimming dead animals, calculating summary statistics, and
generating normalized statistics.
}
\details{
This function executes the following steps:
\enumerate{
\item Initializes an \code{ExperimentData} object containing metadata for the batch.
\item Writes the loading metadata to a CSV file.
\item Generates activity and sleep plots based on the input data.
\item Trims out animals identified as dead using the \code{aliveVsDead} and \code{manualDeadRemoval} functions.
\item Creates a final summary of the bout data and latency statistics.
\item Optionally, generates concatenated genotype plots.
\item Computes summary statistics and normalized statistics for sleep time for all groups.
\item Writes relevant output files, including the final summary and normalized statistics.
}
}
\keyword{Internal}
