% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOneBatch.R
\name{runOneBatch}
\alias{runOneBatch}
\title{Run the Complete Single Batch Analysis (Export)}
\usage{
runOneBatch(oneBatch, control, num_days, font = "plain")
}
\arguments{
\item{oneBatch}{A character string of the Batch folder to be analyzed.}

\item{control}{A character string specifying the control.}

\item{num_days}{A numerical value specifying the number of days to be used in analysis.}

\item{font}{A string variable determining the font style of the produced plots.}
}
\value{
This function does not return a value but performs a series of steps to process the data,
generate plots, and calculate statistics.
}
\description{
This function processes data for a single batch by performing various steps including data cleaning,
generating activity and sleep plots, trimming dead animals, calculating summary statistics, and
generating normalized statistics.
}
\keyword{export}
