% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOneBatch.R
\name{runOneBatch}
\alias{runOneBatch}
\title{Run the Complete Batch Analysis (Internal)}
\usage{
runOneBatch(info, divisions, num_days, pref, controlgeno, controltreat)
}
\arguments{
\item{info}{A data.table containing experimental information (e.g., monitor, genotype, etc.).}

\item{divisions}{A list of time divisions for the analysis.}

\item{num_days}{The number of days to consider for the analysis.}

\item{pref}{A vector of preferences for generating specific plots (e.g., whether to generate concatenated plots).}

\item{controlgeno}{A character string specifying the control genotype.}

\item{controltreat}{A character string specifying the control treatment.}
}
\value{
This function does not return a value but performs a series of steps to process the data,
generate plots, and calculate statistics.
}
\description{
This function processes data for a single batch by performing various steps including data cleaning,
generating activity and sleep plots, trimming dead animals, calculating summary statistics, and
generating normalized statistics.
}
\details{
This function executes the following steps:
\enumerate{
\item Initializes an \code{ExperimentData} object containing metadata for the batch.
\item Writes the loading metadata to a CSV file.
\item Generates activity and sleep plots based on the input data.
\item Trims out animals identified as dead using the \code{aliveVsDead} and \code{manualDeadRemoval} functions.
\item Creates a final summary of the bout data and latency statistics.
\item Optionally, generates concatenated genotype plots.
\item Computes summary statistics and normalized statistics for sleep time for all groups.
\item Writes relevant output files, including the final summary and normalized statistics.
}
}
\keyword{internal}
